update about & conect
